CSNode関連のインスタンス変数、メソッドをMapObjectに移動
